backend/requirements-neoai.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Frontend
gradio==4.15.0

# AI & ML
langchain==0.1.0
langchain-core==0.1.0
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
huggingface-hub==0.19.0

# Database
supabase==2.0.0
python-dotenv==1.0.0

# Utilities
requests==2.31.0
pydantic==2.4.2
pydantic-settings==2.1.0

# System
psutil==5.9.6
pynvml==11.535.108

# Optional: For 4-bit quantization
bitsandbytes==0.41.0

# Logging & Monitoring
python-json-logger==2.0.7
